Refactored: Moved Navbar to a separate component, fixed Link nesting, added TODO for pagination, moved Blog type to types.d.ts
