+++ src/app/page.tsx
@@ -2,19 +2,14 @@
 
 import Link from "next/link";
 import { useState, useEffect } from "react";
-
-type Blog = {
-  id: number;
-  title: string;
-  content: string;
-  createdAt: string;
-};
+import Navbar from "@/components /Navbar";
 
 export default function HomePage() {
   const [blogs, setBlogs] = useState<Blog[]>([]);
 
   useEffect(() => {
     const fetchBlogs = async () => {
+      /////TODO: Add pagination (UI + backend)////
       const res = await fetch("/api/blogs");
       const data = await res.json();
       setBlogs(data.blogs);
@@ -24,30 +19,18 @@ export default function HomePage() {
 
   return (
     <div>
-      {/* Navbar */}
-      <nav className="flex justify-between items-center p-4 bg-white shadow-md sticky top-0 z-50">
-        {/* ////TODO: ADD HOVER ANIMATION & REDIRECT  ///// */}
-        <h1 className="text-xl font-bold text-teal-600">My Blog Web</h1>
-        <div className="space-x-4">
-          <Link href="/" className="text-teal-700 hover:underline">
-            Home
-          </Link>
-          <Link href="/create" className="text-teal-700 hover:underline">
-            Create
-          </Link>
-        </div>
-      </nav>
-
+      <Navbar />
       {/* Hero Section */}
       <section className="h-screen flex flex-col justify-center items-center bg-teal-100 text-center px-4">
         <h2 className="text-4xl font-bold text-teal-800 mb-4">My Blog Web</h2>
         <p className="text-lg text-gray-700 mb-6">
           Your Digital Diary — Capture Your Thoughts.
         </p>
-        <Link href="/create">
-          <button className="bg-teal-600 text-white px-6 py-2 rounded hover:bg-teal-700 transition">
-            + Create a Blog
-          </button>
+        <Link
+          href="/create"
+          className="bg-teal-600 text-white px-6 py-2 rounded hover:bg-teal-700 transition"
+        >
+          + Create a Blog
         </Link>
       </section>
