+++ src/type.d.ts
@@ -1,0 +1,6 @@
+type Blog = {
+  id: number;
+  title: string;
+  content: string;
+  createdAt: string;
+};
